fix lint
change uint32 to int64 to avoid overflow

+++ reedsolomon/reedsolomoncode.go
@@ -2,7 +2,7 @@ package reedsolomon
 
 // Encoding represents an encoded part of a message containing its byte value and its index in the encoded message
 type Encoding struct {
-	Idx int
+	Idx int64
 	Val []byte
 }
